Enhancement: Implement Pixel Art Effect Through Upscaling Rendered Sprite
In main.py, the .Surface function is utilized to render our sprite on a half-resolution display. Subsequently, this rendered sprite is upscaled to fit the larger resolution screen. This deliberate approach results in the enlargement of the sprite, effectively creating a pixel art effect. This enhancement adds visual interest and character to the game's graphics, contributing to its aesthetic appeal.